safari-600.7-branch/LayoutTests/media/progress-events-generated-correctly-expected.txt

Test that progress events are generated during loading of media resource.

EVENT(canplaythrough)
EVENT(playing)
EVENT(ended)
EXPECTED (progressCount > '1') OK
END OF TEST
